P10: stage2 smoke-run passes -- CC.PRG compiles on-target to byte-identical objects
Self-hosted CC.PRG now produces byte-identical ELF objects to the cross-compiler
for a trivial function, a hand-written .s, and a multi-feature sample. Fixes three
self-host-only miscompiles (host c68k.exe is MSVC-built, so only the generated 68k
code was affected):

- parse.c: struct-return-by-value ABI double-count. A leftover x86-64 rule added a
  hidden return-buffer param for aggregates >16B while the codegen already reserves
  the hidden 8(a6) slot uniformly (+4 in assign_lvar_offsets, hard-coded 8(a6) in
  ND_RETURN), shifting every real parameter +4. The assembler's parse_ea returns a
  40-byte EA by value, so every operand parsed as empty -> all abs.L relocations
  (4x-oversized objects).

- codegen68k.c: load_imm printed an int64_t via %ld (a 32-bit long); on the
  big-endian LP32 target that reads the wrong half, so a literal 1 compiled to
  `moveq #0`. Cast the arg to (long) (load_imm only handles 32-bit values).

- libc/core/libc.c: printf never parsed the '+' (or ' ') flag, so the codegen's
  "%s%+ld" relocation-addend format leaked the string literally into the assembly.
  Added +/space sign flags to _vformat.

Also part of enabling the smoke-run:
- osiris_sys.a68: crt0 build_argv parses the PSP command tail into argv[] so
  `CC -c X.C -o X.O` receives its arguments (all c68k Osiris programs; printftest
  unregressed).
- main.c: case-insensitive file-extension detection (Osiris/CP/M filesystems are
  case-folding), plus a native .s->.o assemble path (FILE_ASM) that isolates the
  integrated assembler.
- tools/osiris/smoke-cc.ps1: new stage2 smoke harness (stages CC.PRG + source on a
  FAT12 floppy, runs under sim68k, extracts the object, byte-compares to the host).

+++ src/codegen68k.c
@@ -83,10 +83,15 @@ static void pop64(char *hi, char *lo) {
 
 // Materialize an integer constant into D0 (MOVEQ when it fits a signed byte).
 static void load_imm(int64_t val) {
+  // NB: cast to (long) so the varargs width matches the "%ld" conversion.
+  // int64_t is "long long" here; passing it to a "%ld" (32-bit long) slot
+  // reads the wrong half on a big-endian LP32 self-host (68k) -- e.g. "1"
+  // would print as "0". This function only handles 32-bit-representable
+  // values (64-bit constants go through load_imm64), so (long) is exact.
   if (val >= -128 && val <= 127)
-    println("  moveq #%ld,d0", val);
+    println("  moveq #%ld,d0", (long)val);
   else
-    println("  move.l #%ld,d0", val);
+    println("  move.l #%ld,d0", (long)val);
 }
 
 // Materialize a 64-bit integer constant into D0:D1 (D0=high, D1=low).
